PHP-CS-Fixer fixes

+++ app/code/core/Mage/Payment/Model/Method/Cc.php
@@ -365,9 +365,9 @@ class Mage_Payment_Model_Method_Cc extends Mage_Payment_Model_Method_Abstract
     {
         $info = $this->getInfoInstance();
         if ($this->_isPlaceOrder()) {
-            return (double)$info->getOrder()->getQuoteBaseGrandTotal();
+            return (float)$info->getOrder()->getQuoteBaseGrandTotal();
         } else {
-            return (double)$info->getQuote()->getBaseGrandTotal();
+            return (float)$info->getQuote()->getBaseGrandTotal();
         }
     }
